requirements.txt
streamlit
numpy
Pillow
tensorflow
scikit-image